Fix compilation under Mono.
The problem is that we had e.g.

	namespace DbLinq.Vendor.Implementation {
	#if !MONO_STRICT
		public
	#endif
		partial class SchemaLoader {
	#if !MONO_STRICT
			public
	#endif
			class DataName : IDataName {/*...*/}
		}
	}
	namespace DbLinq.Oracle {
	#if !MONO_STRICT
		public
	#endif
		class OracleSchemaLoader : SchemaLoader {
			protected class StoredProcedureParameter {
				public DataType Type {get; set;}
			}
		}
	}

The problem is that when MONO_STRICT is set, SchemaLoader.DataName is now a
_private_ type, which is thus inaccessible to OracleSchemaLoader, resulting in
a compilation error.  Oops.

Furthermore, the '#if !MONO_STRICT' use is actually pointless for nested
types, as their accessibility can be no more public than their enclosing type
(SchemaLoader), which would be internal.  So there's no problem in just
leaving the nested types 'public' (and this reduces #if usage to boot!).