Permit use of generic queries.
Reported at:
http://groups.google.com/group/dblinq/browse_frm/thread/2855c41f7e907e7d/7e5b6fc687d6eed2?lnk=gst&q=performing+common+operations+on+two+different#7e5b6fc687d6eed2

The issue is that in a "generic" query:

	static IEnumerable<T> MatchAddress<T>(IQueryable<T> q)
		where T : IHasAddress
	{
		return q.OrderByDescending(v => v.Address.Length);
	}

The compiler turns the 'v => v.Address.Length' expression tree into one
containing a "Convert" node; in VS Debugging-ese:

	v => Convert(v).Address.Length

This "spurious" Convert() node blew things up, as we're not actually doing any
converting here (UnaryExpression.Method is null), yet when we hit the SQL
generation phase DbLinq wanted to conver to...something, which it couldn't.

In this case, remove the Convert() node and return the underlying
TableExpression when we notice this condition.